claude_and_gemini/pyproject.toml
[tool.poetry]
name = "github-tidy"
version = "0.1.0"
description = "A tool to manage GitHub branches by archiving inactive ones"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "github_branch_manager", from = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
PyGithub = "^2.1.1"
python-dotenv = "^1.0.0"
slack-sdk = "^3.21.3"
google-cloud-logging = "^3.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
github-tidy = "github_branch_manager.main:main" 